wip: move models into their own package
